requirements.txt
pydantic==2.8.2
opencv-python-headless==4.10.0.84
firebase-admin==6.5.0
pinecone==5.0.1
fastapi==0.112.1
uvicorn==0.30.6
numpy==1.26.4
python-dotenv==1.0.1
python-multipart==0.0.9
loguru==0.7.2
uvicorn==0.30.6
aiofiles~=24.1.0
torch~=2.5.0
torchvision~=0.20.0
pydantic-settings==2.6.1
pyinstrument==5.0.0
APScheduler==3.10.4
requests==2.31.0
slowapi==0.1.9
sentry-sdk==2.19.2